refactor: move client utils into their own file
